% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{average_buckets}
\alias{average_buckets}
\title{Averages a vector based on bucket sizes}
\usage{
average_buckets(a, buckets)
}
\arguments{
\item{a}{the vector to be bucketed and averaged}

\item{buckets}{the vector of bucket sizes to average a over}
}
\value{
the vector of averaged a
}
\description{
Given a vector (a) and another vector of bucket sizes, returns the averaged vector (a)
}

