% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{b.calc}
\alias{b.calc}
\title{Bite rate calculation}
\usage{
b.calc(params, R0)
}
\arguments{
\item{params}{Vector of parameters matching those returned by \code{\link{setupListPars}}}

\item{R0}{desired R0 value}
}
\value{
A single value for bite rate
}
\description{
Calculates the bite rate needed to generate a given R0 value, assuming that all other parameters are fixed.
}
\seealso{
\code{\link{r0.calc}}
}

