% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_alphas}
\alias{calculate_alphas}
\title{Converts to linear scale}
\usage{
calculate_alphas(y, probMicro, sampFreq)
}
\arguments{
\item{x}{the double to be converted back to linear scale}

\item{min}{the minimum value on the linear scale}

\item{max}{the maximum value on the linear scale}
}
\value{
the value converted to a linear scale
}
\description{
Converts to linear scale
}

