% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_alphas_prob_buckets}
\alias{calculate_alphas_prob_buckets}
\title{Calculates probabilities of microcephaly over time for given time bucket sizes}
\usage{
calculate_alphas_prob_buckets(y, probMicro, baselineProb, times)
}
\arguments{
\item{y}{the matrix of pregnant adult counts. First column should be times}

\item{probMicro}{probability of developing microcephaly given infection}

\item{baselineProb}{the baseline probability of giving birth to an infant with microcephaly}

\item{times}{matrix of times to create alphas over. First column is start of bucket, last column is end of bucket}
}
\value{
the vector of alphas
}
\description{
Calculates probabilities of microcephaly over time for given time bucket sizes
}

