% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_alphas_prob_sampfreq}
\alias{calculate_alphas_prob_sampfreq}
\title{Calculates probability of microcephaly over time with regular sampling intervals}
\usage{
calculate_alphas_prob_sampfreq(y, probMicro, baselineProb, sampFreq)
}
\arguments{
\item{y}{the matrix of pregnant adult counts. First column should be times}

\item{probMicro}{probability of developing microcephaly given infection}

\item{baselineProb}{the baseline probability of giving birth to an infant with microcephaly}

\item{sampFreq}{the frequency of sampling}
}
\value{
the value converted to a linear scale
}
\description{
Calculates probability of microcephaly over time with regular sampling intervals
}

