% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{calculate_AR}
\alias{calculate_AR}
\title{Attack rate}
\usage{
calculate_AR(r0)
}
\arguments{
\item{R0}{the value of R0 to test}
}
\value{
the estimated attack rate
}
\description{
Calculates attack rate using final size equation and R0
}

