% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{calculate_gamma_mode}
\alias{calculate_gamma_mode}
\title{Gamma mode}
\usage{
calculate_gamma_mode(mean, var)
}
\arguments{
\item{mean}{the gamma mean}

\item{var}{the gamma variance}
}
\value{
the gamma mode
}
\description{
Calculates the mode of a gamma distribution given the mean and variance
}

