% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{combine_chains}
\alias{combine_chains}
\title{Combine MCMC chains}
\usage{
combine_chains(chains)
}
\arguments{
\item{chains}{the list of chains}
}
\value{
a single data frame
}
\description{
Given a list of MCMC chains, rbinds them all to give one big chain
}

