% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_data.R
\name{convert_name_to_location}
\alias{convert_name_to_location}
\title{Location conversion}
\usage{
convert_name_to_location(name)
}
\arguments{
\item{name}{the vector or scalar of lower case location names}
}
\value{
the same vector but with actual strings
}
\description{
Converts lower-case simplified names to actual strings
}

