% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_data.R
\name{convert_name_to_location_factor}
\alias{convert_name_to_location_factor}
\title{Convert name to factor}
\usage{
convert_name_to_location_factor(locations)
}
\arguments{
\item{locations}{the vector of lower case location names to convert}
}
\value{
the converted locations with proper naming and factored
}
\description{
Function to convert location strings to factors (with correct accents etc)
}

