% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{convert_name_to_state}
\alias{convert_name_to_state}
\title{State conversion}
\usage{
convert_name_to_state(name)
}
\arguments{
\item{name}{the vector or scalar of lower case state names}
}
\value{
the same vector but with actual strings
}
\description{
Contains a vector of the brazilian state names to convert lower-case simplified names to actual strings
}

