% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{convert_name_to_state_factor}
\alias{convert_name_to_state_factor}
\title{Convert name to factor}
\usage{
convert_name_to_state_factor(states)
}
\arguments{
\item{states}{the vector of lower case state names to convert}
}
\value{
the converted states with proper naming and factored
}
\description{
Function to convert state strings to factors (with correct accents etc)
}

