% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{createCounts}
\alias{createCounts}
\title{Create count table}
\usage{
createCounts(dat)
}
\arguments{
\item{dat}{the matrix to be counted}
}
\value{
a data frame of counted data
}
\description{
Given a matrix of data points (rows = times, columns = individuals), returns the summarised count data
}

