% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_setup.R
\name{createLocationParTable}
\alias{createLocationParTable}
\title{MCMC location-specific parameter table creation}
\usage{
createLocationParTable(locationInfo, saveFile = NULL)
}
\arguments{
\item{locationInfo}{data frame of locations - with L_H, N_H and rawName}

\item{saveFile}{if provided, writes the parameter table to the given filename}
}
\value{
a matrix of needed settings for the MCMC algorithm. For each parameter, gives a name, lower and upper bounds, initial step sizes and whether or not the parameter should be fixed.
}
\description{
Sets up the parameter table for all locations for use in \code{\link{run_MCMC}}
}

