% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{createParTable}
\alias{createParTable}
\title{MCMC parameter table creation}
\usage{
createParTable(version = 1, realDat = NULL, saveFile = NULL)
}
\arguments{
\item{version}{the version of the model to use. See \code{\link{print_version_names}}}

\item{realDat}{if real data is available, adds parameters for the states given}

\item{saveFile}{if provided, writes the parameter table to the given filename}
}
\value{
a matrix of needed settings for the MCMC algorithm. For each parameter, gives a name, lower and upper bounds, initial step sizes and whether or not the parameter should be fixed.
}
\description{
Sets up the parameter table for use in \code{\link{run_metropolis_MCMC}}
}

