% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{createStateParTable}
\alias{createStateParTable}
\title{MCMC state-specific parameter table creation}
\usage{
createStateParTable(stateDat, saveFile = NULL)
}
\arguments{
\item{stateDat}{data frame of states - had L_H, N_H and local name}

\item{saveFile}{if provided, writes the parameter table to the given filename}
}
\value{
a matrix of needed settings for the MCMC algorithm. For each parameter, gives a name, lower and upper bounds, initial step sizes and whether or not the parameter should be fixed.
}
\description{
Sets up the parameter table for all states for use in \code{\link{run_metropolis_MCMC}}
}

