% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{cumul_true}
\alias{cumul_true}
\title{Cumulative true}
\usage{
cumul_true(x)
}
\arguments{
\item{x}{the vector to be tested}
}
\value{
a vector counting number of successive TRUEs
}
\description{
Gets number of consecutive values in a vector that are TRUE
}

