% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{density.calc}
\alias{density.calc}
\title{Density calculation}
\usage{
\method{density}{calc}(params, R0)
}
\arguments{
\item{params}{Vector of parameters matching those returned by \code{\link{setupListPars}}}

\item{R0}{desired R0 value}
}
\value{
A single value for mosquito density
}
\description{
Calculates the density needed to generate a given R0 value, assuming that all other parameters are fixed.
}
\seealso{
\code{\link{r0.calc}}
}

