% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{density_plot_func}
\alias{density_plot_func}
\title{Microcephaly parameter density plot}
\usage{
density_plot_func(varName, dat, xlabtext = "", weeks = 1, xlim = NULL,
  xbreaks = NULL)
}
\arguments{
\item{varName}{the name of the variable to plot}

\item{dat}{the melted MCMC chain}

\item{xlabtext}{optional x-axis title}

\item{weeks}{if set, this divides everything by this amount. ie. if we want value in weeks rather than days}

\item{xlim}{optional vector of lower and upper x axis limits}

\item{xbreaks}{optional vector of x axis break points}
}
\value{
a ggplot object density plot
}
\description{
Given a melted MCMC chain with a column for variable name, plots the posterior density with shaded regions for 95% CI.
}

