% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{density_plot_state}
\alias{density_plot_state}
\title{Posterior density plot}
\usage{
density_plot_state(meltedChain, variableName, xlimits = NULL, title = "",
  breaks = NULL, labels = NULL)
}
\arguments{
\item{meltedChain}{the melted MCMC chain}

\item{variableName}{the variable to plot}

\item{xlimits}{optional vector of lower and upper x axis range}

\item{title}{the title to give to the plot}

\item{breaks}{option x axis breaks}

\item{optional}{x axis labels. Must match breaks length}
}
\value{
a ggplot object of posterior density
}
\description{
Plots posterior density of a given variable, facetted by state
}

