% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{exampleParTab}
\alias{exampleParTab}
\title{Example parameter table}
\format{An object of class \code{data.frame} with 19 rows and 9 columns.}
\usage{
data(exampleParTab)
}
\description{
Example parameter table
}
\keyword{datasets}

