% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{extra_microceph_calculations}
\alias{extra_microceph_calculations}
\title{Post calculations}
\usage{
extra_microceph_calculations(chain, version = 1, microceph_limit = 0.001,
  runs = 100)
}
\arguments{
\item{chain}{the full MCMC chain to do post analysis on}

\item{version}{the version of the model run here}

\item{microceph_limit}{the probability of microcephaly given infection to use as the cut off}

\item{runs}{the number of samples to take}
}
\value{
an MCMC chain of the added analyses
}
\description{
Adds some parameter estimates for peripheral microcephaly risk parameters eg. mode. Note that this only applies to version 1 of the model; if not version 1, will return lots of zeros
}

