% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{find_peak_times}
\alias{find_peak_times}
\title{Find peak times}
\usage{
find_peak_times(parTab, ts = seq(0, 3003, by = 1), solver = "rlsoda")
}
\arguments{
\item{parTab}{table of parameters with names, values and corresponding state}

\item{ts}{vector of time in days in solve model over}

\item{solver}{which ODE solver to use}
}
\value{
a vector of peak times for each state in the parameter table
}
\description{
Given a table of parameters as returned by \code{\link{setupParTable}}, finds the time of peak ZIKV incidence for each state
}

