% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{forecast_CRS}
\alias{forecast_CRS}
\title{Model function for CRS prediction}
\usage{
forecast_CRS(pars, startDays, endDays, buckets, microCeph, births, inc, nh,
  inc_buckets, inc_start, inc_end, perCap = FALSE)
}
\description{
Using the microcephaly risk model, produces a forecast of CRS affected births for all time periods and calculates a likelihood of observing given CRS cases. Uses reported incidence as per capita risk (after dividing my reporting rate).
}

