% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{forecast_microceph}
\alias{forecast_microceph}
\title{Microcephaly incidence forecast}
\usage{
forecast_microceph(chain, microParChain = NULL, parTab, ts, runs,
  origin = "2013-01-01")
}
\arguments{
\item{chain}{the MCMC chain to simulate from}

\item{microParChain}{the MCMC chain for microcephaly parameters}

\item{parTab}{the parameter table matching the generated MCMC chain}

\item{ts}{the vector of times to solve the model over}

\item{runs}{number of draws to make to get uncertainty}

\item{origin}{the date from which to forecast from}
}
\value{
a table of incidence and dates
}
\description{
Given an MCMC chain, generates the incidence of microcephaly affected births
}

