% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{forecast_microceph_indiv}
\alias{forecast_microceph_indiv}
\title{Microcephaly forecast single pars}
\usage{
forecast_microceph_indiv(pars, local, parTab, ts, weeks = FALSE)
}
\arguments{
\item{pars}{the model parameters to solve}

\item{local}{the name of the Brazilian state}

\item{parTab}{the parameter table to draw names etc. from}

\item{ts}{the vector of times to solve the model over. Should be in days, but will be converted to months.}

\item{weeks}{if preferred, will return weekly proportions rather than months}
}
\value{
a data frame of months and proportion microcephaly affected births
}
\description{
Forecasts monthly microcephaly birth proportion for a single set of model parameters
}

