% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{generate_allowable_params}
\alias{generate_allowable_params}
\title{Generate allowable starting parameters}
\usage{
generate_allowable_params(peakTimings = 927, peakTimeRange = 60, stateNames,
  parTab = NULL, allowableParsFile = "allowablePars.csv", R0max = 6.5)
}
\arguments{
\item{peakTimeRange}{the desired allowable peak time range (ie. square prior)}

\item{stateNames}{the set of Brazilian states for which allowable parameters should be generated. Must match those in the microDatFile}

\item{parTab}{the parameter table as returned by \code{\link{setupParTable}}}

\item{allowableParsFile}{file location for the allowable parameters table, if it exists.}

\item{peakTime}{the desired central peak time. Defaults to 927}

\item{microDatFile}{the location of the microcephaly data file which is needed to obtain N_H and L_H for each state. Note that this is only needed if no parTab is provided}
}
\value{
a table of allowable parameters with columns for t0, mosquito density (R0), corresponding state (as this will vary by N_H and life expectancy), and corresponding peak time
}
\description{
Given a desirable SEIR peak time and time window, generates a set of R0 and t0 parameters that satisfy this peak time
}

