% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_foi}
\alias{generate_foi}
\title{FOI calculation}
\usage{
generate_foi(IM, NH, b, pMH, tstep)
}
\arguments{
\item{IM}{a numeric vector of number of infected mosquitoes over time}

\item{NH}{the constant human population size}

\item{b}{the per vector per day bite rate}

\item{pMH}{the probability of transmission upon bite}

\item{the}{time step over which to bucket the force of infection. Best left to 1 for 1 day}
}
\value{
the vector of FOI
}
\description{
Calculates the force of infection over time for an SEIR model
}

