% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_incidence_data}
\alias{generate_incidence_data}
\title{Simulation multi-state Zika incidence data}
\usage{
generate_incidence_data(t_pars, paramTable, dataRange = c(0, t_pars["dur"]),
  weeks = TRUE, noise = TRUE, peakTimeRange = NULL)
}
\arguments{
\item{t_pars}{vector of two components - the total run time in days and the time step for the ODE solver}

\item{paramTable}{Table of all parameters needed to solve the SEIR model and microcephaly simulation for each unique state. See \code{\link{setupParTable}}}

\item{dataRange}{a vector with lower and upper bounds on days for simulation data}

\item{weeks}{Defaults to TRUE. If true. returns weekly microcephaly births rather than by month.}

\item{noise}{bool indicating if noise should be added}

\item{peakTimeRange}{if not NULL, returns an additional data frame with a range around the actual incidence peak time}

\item{pars}{a named vector with all of the necessary parameters to solve the ODE model. See \code{\link{setupParsODE}}}
}
\value{
a data frame of incidence data and corresponding times.
}
\description{
Given all of the parameters needed to solve the SEIR model, produces a data frame of simulated incidence data for multiple states. Produces data for each unique state in the paramTable argument.
}

