% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generate_multiple_data}
\alias{generate_multiple_data}
\title{Simulation multi-state microcephaly data}
\usage{
generate_multiple_data(ts = seq(0, 3003, by = 1), paramTable, weeks = FALSE,
  dataRangeMicro, dataRangeInc, noise = NULL, peakTimeRange = NULL)
}
\arguments{
\item{paramTable}{Table of all parameters needed to solve the SEIR model and microcephaly simulation for each unique state. See \code{\link{setupParTable}}}

\item{weeks}{Defaults to FALSE. If true. returns weekly microcephaly births rather than by month.}

\item{dataRangeMicro}{a vector with lower and upper bounds on days for simulated microcephaly data}

\item{dataRangeInc}{a vector with lower and upper bounds on days for simulated ZIKV incidence data}

\item{noise}{bool indicating if noise should be added to the data}

\item{peakTimeRange}{the allowable prior range for ZIKV incidence peak time}

\item{t_pars}{vector of two components - the total run time in days and the time step for the ODE solver}
}
\value{
a data frame of microcephaly births, total births and corresponding times.
}
\description{
Given all of the parameters needed to solve the SEIR model, produces a data frame of simulated microcephaly affected births for multiple states. Produces data for each unique state in the paramTable argument.
}

