% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_plot_date_labels}
\alias{generate_plot_date_labels}
\title{Generate date labels}
\usage{
generate_plot_date_labels(birth_dat)
}
\arguments{
\item{birth_dat}{the data frame of microcephaly data}
}
\value{
the vector of labels
}
\description{
Given a data frame of microcephaly data, generates a vector of x-labels that represent dates
}

