% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_probM}
\alias{generate_probM}
\title{Time varying microcephaly risk}
\usage{
generate_probM(IM, NH, probM, b, pMH, bp, tstep)
}
\arguments{
\item{IM}{a numeric vector of number of infected mosquitoes over time}

\item{NH}{the constant human population size}

\item{probM}{the time varying risk of developing microcephaly given infection over the course of gestation}

\item{b}{the per vector per day bite rate}

\item{pMH}{the probability of transmission upon bite}

\item{bp}{the baseline probability of microcephaly}

\item{tstep}{the time step for the buckets. Best left to 1.}
}
\value{
the vector of microcephaly probabilities
}
\description{
Given the epidemic model parameters and number of infected mosquitoes over time, calculates the probability of a birth having microcephaly at a given time
}

