% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_probM_aux}
\alias{generate_probM_aux}
\title{Time varying microcephaly risk AUX}
\usage{
generate_probM_aux(riskI, probM, bp)
}
\arguments{
\item{riskI}{the relative risk of infection over time as generated by \link{\code{generate_riskI}}}

\item{probM}{the time varying risk of developing microcephaly given infection over the course of gestation. Time steps should be the same as riskI}

\item{bp}{the baseline probability of microcephaly}
}
\value{
the vector of microcephaly probabilities
}
\description{
Calculates the probability of a birth having microcephaly at a given time
}

