% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_probM_forecast_NEW}
\alias{generate_probM_forecast_NEW}
\title{Time varying microcephaly risk not aborted}
\usage{
generate_probM_forecast_NEW(riskI, probM, bp, abortion_rate, birth_reduction,
  switch_t, abortion_time, ABORTED)
}
\arguments{
\item{riskI}{the relative risk of infection over time as generated by generate_riskI}

\item{probM}{the time varying risk of developing microcephaly given infection over the course of gestation. Time steps should be the same as riskI}

\item{bp}{the baseline probability of microcephaly}

\item{abortion_rate}{probability of a women getting an abortion given that she was infected}

\item{birth_reduction}{reduced probability of getting infected, synonymous with probability of avoiding pregnancy}

\item{switch_t}{integer for index after which behaviour changes}

\item{abortion_time}{integer for last allowable abortion day/week}

\item{ABORTED}{if TRUE, then return the number of aborted cases. If FALSE, return the number of not aborted cases.}
}
\value{
the vector of microcephaly probabilities
}
\description{
Calculates the probability of a birth having microcephaly at a given time
}

