% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_riskI}
\alias{generate_riskI}
\title{Risk of infection}
\usage{
generate_riskI(foi, riskS, tstep)
}
\arguments{
\item{foi}{the time varying force of infection as generated by \link{\code{generate_foi}}}

\item{riskS}{the cumulative probability of remaining susceptible over the course of the epidemic as generated by \link{\code{generate_riskS}}}

\item{tstep}{the time step for the buckets. Best left to 1.}
}
\value{
the vector of relative infection risks
}
\description{
Calculates the relative risk of infection over time
}

