% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_riskS}
\alias{generate_riskS}
\title{Cumulative susceptible probability}
\usage{
generate_riskS(foi, tstep)
}
\arguments{
\item{foi}{the time varying force of infection as generated by generate_foi}

\item{tstep}{the time step for the buckets. Best left to 1.}
}
\value{
the vector of cumulative escape probabilities
}
\description{
Calculates the cumulative probability of remaining susceptible over the course of the epidemic
}

