% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_y0s}
\alias{generate_y0s}
\title{Generates y0s}
\usage{
generate_y0s(N_H, density, iniI = 10)
}
\arguments{
\item{N_H}{human population size}

\item{density}{number of mosquitoes per person}
}
\value{
a vector of initial population sizes
}
\description{
Generates initial values for the simple SEIR model given population size and mosquito density
}

