% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_setup.R
\name{generateStartingParTab}
\alias{generateStartingParTab}
\title{MCMC start parTab}
\usage{
generateStartingParTab(parTab, peakTimes = NULL, restrictedR0 = TRUE,
  allowableParsFile = "", allowableStarts = NULL)
}
\arguments{
\item{parTab}{the parameter table to generate starting positions from}

\item{peakTimes}{the ODE solver and likelihood function is a bit tempermental to the choice of starting parameters for R0 and the epidemic seed time, t0.}

\item{restrictedR0}{if TRUE, sets starting points for density and t0 based on generated allowable values}

\item{allowableParsFile}{set to "" if not available. These allowable start points take a while to generate (\code{\link{generateAllowableParams}}), so it can sometimes be helpful to pre-compute these and then pass the file location}

\item{allowableStarts}{the pre-computed matrix of allowable parameters}
}
\value{
a parameter table in the same form as that which was passed in, but with randomised values
}
\description{
Generates random starting points from lower_bounds and upper_bounds to seed the MCMC chain for free parameters
}

