% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{get_correct_order}
\alias{get_correct_order}
\title{State order}
\usage{
get_correct_order(named = TRUE, normalised = FALSE)
}
\arguments{
\item{named}{if TRUE, converts the state names to full names}

\item{normalised}{if TRUE, orders by per birth microcephaly rate}

\item{birth_dat}{the data frame of data}
}
\value{
a vector of ordered state names
}
\description{
Gets the correct state order for all plots, sorted by number of microcephaly cases
}

