% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{get_epidemic_states}
\alias{get_epidemic_states}
\title{Subset epidemic states}
\usage{
get_epidemic_states(dat, micro = TRUE, lim = 575)
}
\arguments{
\item{dat}{the data frame of data to look through (microcephaly or incidence)}

\item{micro}{if TRUE, looks at microcephaly data. Incidence data otherwise}

\item{lim}{in days, the start of the period of consideration}
}
\value{
a list with a vector of epidemic and non-epidemic states
}
\description{
Gets those states that fulfill the epidemic criteria of 2sd above mean since July for 3 months
}

