% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{get_max_micro}
\alias{get_max_micro}
\title{Max micro risk}
\usage{
get_max_micro(chain, scale)
}
\arguments{
\item{chain}{the MCMC chain to run over}

\item{scale}{scales microcephaly risk curve by assumed reporting proportion}
}
\value{
a vector of estimated maximum risk from the MCMC chain
}
\description{
For a given MCMC chain, calculates the maximum risk of microcephaly given infection
}

