% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{get_max_micro_day}
\alias{get_max_micro_day}
\title{Max micro risk day}
\usage{
get_max_micro_day(chain)
}
\arguments{
\item{chain}{the MCMC chain to run over}
}
\value{
a vector of estimated maximum risk days from the MCMC chain
}
\description{
For a given MCMC chain, calculates the day of maximum risk of microcephaly given infection
}

