% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{get_microceph_range}
\alias{get_microceph_range}
\title{Microcephaly risk range}
\usage{
get_microceph_range(chain, location = NULL, runs, limit = 0.001,
  scale = 1)
}
\arguments{
\item{chain}{the MCMC chain}

\item{location}{the given location to subset by, if present.}

\item{runs}{number of samples to take}

\item{limit}{the risk limit to consider as being at risk}

\item{scale}{scales microcephaly risk curve by assumed reporting proportion}
}
\value{
a list with three vectors. 1: first day of significant risk, 2: last day of significant risk; 3: the number of days spent at risk
}
\description{
Getting microcephaly risk curve range from a chain for a given location
}

