% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_state_name}
\alias{get_state_name}
\title{Translates state code to name}
\usage{
get_state_name(state)
}
\arguments{
\item{state}{the state code to be changed}
}
\value{
a string
}
\description{
As described
}

