% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{get_states}
\alias{get_states}
\title{Get states lowercase}
\usage{
get_states()
}
\value{
vector of all states
}
\description{
Gives a vector of all states in Brazil. Names are actual names, values are lower case
}

