% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{get_upper_micro_bound}
\alias{get_upper_micro_bound}
\title{Upper microbound}
\usage{
get_upper_micro_bound(chain, limit = 0.001, scale = 1)
}
\arguments{
\item{chain}{the MCMC chain to run over}

\item{limit}{the probability of microcephaly considered at risk}

\item{scale}{scales microcephaly risk curve by assumed reporting proportion}
}
\value{
a vector of last risk days
}
\description{
For a given MCMC chain, calculates the last day at risk
}
\seealso{
\code{\link{get_microceph_range}}
}

