% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{incidence_likelihood}
\alias{incidence_likelihood}
\title{Incidence likelihood (binomial)}
\usage{
incidence_likelihood(perCapInc, inc, N_H)
}
\arguments{
\item{perCapInc}{expected per capita incidence of Zika}

\item{inc}{vector of actual incidence}

\item{N_H}{vector of population sizes (ie. per capita inc)}
}
\value{
a single log likelihood
}
\description{
Given time-varying probabilities of observing Zika incidence and actual data, returns a single likelihood assuming binomially distributed observations
}

