% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{incidence_likelihood_norm}
\alias{incidence_likelihood_norm}
\title{Incidence likelihood normal}
\usage{
incidence_likelihood_norm(perCapInc, inc, N_H, lik_sd)
}
\arguments{
\item{perCapInc}{expected per capita incidence of Zika}

\item{inc}{vector of actual incidence}

\item{N_H}{vector of population sizes (ie. per capita inc)}

\item{lik_sd}{standard deviation of the observation distribution}
}
\value{
a single log likelihood
}
\description{
Given time-varying probabilities of observing Zika incidence and actual data, returns a single likelihood assuming normally distributed observations
}

