% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{isolate_location}
\alias{isolate_location}
\title{Isolate location from chain}
\usage{
isolate_location(local, i, chains, parTab, extraNames = NULL)
}
\arguments{
\item{local}{the name of the location}

\item{i}{the index of the location used for indexing the right column (for example, first location in the table is 0, second is 1 etc...)}

\item{chains}{the MCMC chain to extract from}

\item{parTab}{the parameter table used to generate this MCMC chain. This is needed to ensure that parameter names are correct}

\item{extraNames}{vector of additional names in the MCMC chain}
}
\value{
an MCMC chain for just this location, with R0 and attack rate calculated
}
\description{
Given the name of a location and its corresponding index, returns only those parameters relevant to that location from the MCMC chain
}

