% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{isolate_state}
\alias{isolate_state}
\title{Isolate state from chain}
\usage{
isolate_state(local, i, chains, parTab, extraNames = NULL)
}
\arguments{
\item{local}{the name of the state}

\item{i}{the index of the state used for indexing the right column (for example, first state in the table is 0, second is 1 etc...)}

\item{chains}{the MCMC chain to extract from}

\item{parTab}{the parameter table used to generate this MCMC chain. This is needed to ensure that parameter names are correct}

\item{extraNames}{vector of additional names in the MCMC chain}
}
\value{
an MCMC chain for just this state, with R0 and attack rate calculated
}
\description{
Given the name of a state and its corresponding index, returns only those parameters relevant to that state from the MCMC chain
}

