% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood}
\alias{likelihood}
\title{Calculate gaussian mixture model likelihood}
\usage{
likelihood(dat, alphas, mus, sds)
}
\arguments{
\item{dat}{the matrix of data}

\item{alphas}{matrix of of alphas. 2 columns (1 for both distributions) with number of rows matching number of rows in dat}

\item{mus}{vector of mus for the 2 distributions}

\item{sds}{vector of sds for the 2 distributions}
}
\value{
a single log likelihood
}
\description{
Given a matrix of data points (columns represent individuals, rows represent sampling times), computes the log likelihood using the gaussian mixture model with known alphas, mus and sds.
}

