% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood_prob}
\alias{likelihood_prob}
\title{Calculate likelihood for probability data}
\usage{
likelihood_prob(dat, alphas)
}
\arguments{
\item{dat}{the matrix of data. 2 columns for counts of positive and total births, and N rows for each sampling time}

\item{alphas}{vector of probability of giving birth to an infant with microcephaly over time}
}
\value{
a single log likelihood
}
\description{
Given a matrix of threshold data (ie. microcephaly or not) for different sampling times, gives a log likelihood with the given parameter values
}

