% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood_probM}
\alias{likelihood_probM}
\title{Likelihood function for time-varying microcephaly}
\usage{
likelihood_probM(microBirths, allBirths, probM)
}
\arguments{
\item{microBirths}{the vector of observed microcephaly cases over time}

\item{allBirths}{the corresponding total number of births}

\item{probM}{the corresponding vector of microcephaly probabilities as calculated by generate_probM.}
}
\value{
a single likelihood value
}
\description{
Calculates the likelihood of observing a vector of microcephaly births given the total number of births and microcephaly probabilities. Note that all vectors must be equal lengths.
}

