% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood_probM_all}
\alias{likelihood_probM_all}
\title{Likelihood function for time-varying microcephaly}
\usage{
likelihood_probM_all(microBirths, allBirths, IM, probM, NH, b, pHM, bp, tstep)
}
\arguments{
\item{microBirths}{the vector of observed microcephaly cases over time}

\item{allBirths}{the corresponding total number of births}

\item{IM}{a numeric vector of number of infected mosquitoes over time}

\item{probM}{the time varying risk of developing microcephaly given infection over the course of gestation}

\item{NH}{the constant human population size}

\item{b}{the per vector per day bite rate}

\item{bp}{the baseline probability of microcephaly}

\item{tstep}{the time step for the buckets. Best left to 1.}

\item{pMH}{the probability of transmission upon bite}
}
\value{
a single likelihood value
}
\description{
Calculates the likelihood of observing a vector of microcephaly births given the total number of births and microcephaly probabilities. Takes model generated/used parameters.
}
\seealso{
\link{\code{likelihood_probM}}
}

