% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood_threshold}
\alias{likelihood_threshold}
\title{Calculate likelihood for threshold data}
\usage{
likelihood_threshold(dat, alphas, mus, sds, threshold)
}
\arguments{
\item{dat}{the matrix of data. 2 columns for counts of positive and negative, and N rows for each sampling time}

\item{alphas}{matrix of of alphas. 2 columns (1 for both distributions) with number of rows matching number of rows in dat}

\item{mus}{vector of mus for the 2 distributions}

\item{sds}{vector of sds for the 2 distributions}

\item{threshold}{the threshold value below which a data point is classified as having microcephaly}
}
\value{
a single log likelihood
}
\description{
Given a matrix of threshold data (ie. microcephaly or not) for different sampling times, gives a log likelihood with the given parameter values
}

