% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{load_mcmc_chains}
\alias{load_mcmc_chains}
\title{Load MCMC chains}
\usage{
load_mcmc_chains(location = "", asList = FALSE, convertMCMC = FALSE,
  unfixed = TRUE, thin = 1, burnin = 1e+05, unique_names = TRUE)
}
\arguments{
\item{location}{the path to the directory to read from}

\item{asList}{if TRUE, returns the MCMC chains as a list}

\item{convertMCMC}{if TRUE, converts the chains to mcmc objects (coda)}

\item{unfixed}{if TRUE, only includes columns for estimated parameters (not fixed parameters)}

\item{thin}{thins the read in MCMC chains by the specified amount (useful for speed when testing)}

\item{burnin}{discards the burn in period}

\item{unique_names}{if TRUE, gives column names as if read in by read.csv}
}
\value{
the list or data frame of MCMC chains
}
\description{
From a working directory with MCMC chains (files ending _slice_chain.csv), reads these in and puts them into an MCMC list
}

