% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{locationInfo}
\alias{locationInfo}
\title{Location information table}
\format{A data frame}
\usage{
data(locationInfo)
}
\description{
Table containing information on population size, life expectancy and full location names
}
\keyword{datasets}

