% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{make_every_so_often}
\alias{make_every_so_often}
\title{Function that evaluations to true every nth time it is called}
\usage{
make_every_so_often(iterations = 1)
}
\description{
Function that evaluations to true every nth time it is called
}

