% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{mcmc_diagnostics}
\alias{mcmc_diagnostics}
\title{MCMC diagnostic tests}
\usage{
mcmc_diagnostics(mcmc_chains, filename, param_table, VERBOSE = FALSE)
}
\arguments{
\item{mcmc_chains}{the entire MCMC chain to be tested}

\item{filename}{the full file path at which to save the diagnostics. _gelman.pdf will be appended, for example}

\item{param_table}{the parameter table}

\item{VERBOSE}{boolean flag for additional output. Defaulst to FALSE}
}
\value{
returns any error messages raised during the tests
}
\description{
Runs some basic MCMC diagnostics on the given chain and saves a few plots. The diagnostics are:
\itemize{
\item{Gelman Diagnostics: }{Saves the Gelman plots at the given file location}
\item{Auto-correlation: }{Saves autocorrelation plots at the given file location}
}
}

