% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{mcmc_simple}
\alias{mcmc_simple}
\title{Slice MCMC sampler}
\usage{
mcmc_simple(target, x_init, fixed, nsteps, w, lower = -Inf, upper = -Inf,
  print_every = 1, save_file = NULL)
}
\arguments{
\item{target}{pointer to the posterior (target) function, that should take only one argument - the vector of model parameters}

\item{x_init}{initial paramter values for the target function}

\item{fixed}{a vector of 1/0 (TRUE/FALSE) matching the target function, indicating which parameters should be updated}

\item{nsteps}{number of iterations}

\item{w}{vector of widths for the slice sampler}

\item{lower}{optional vector of lower parameter bounds}

\item{upper}{options vector of upper parameter points}

\item{print_every}{how often should output be printed}
}
\value{
a dataframe of the MCMC chain
}
\description{
Typically uses a univariate slice sampler for proposal update. Performs MCMC on a given target function
}

