% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{median.quantile}
\alias{median.quantile}
\title{Median and quantiles}
\usage{
\method{median}{quantile}(x)
}
\arguments{
\item{x}{the vector}
}
\value{
vector with 95% quantiles and median
}
\description{
Calculates the median and 95% quantiles from a vector
}

