% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{microceph_v1}
\alias{microceph_v1}
\title{Microcephaly risk V1}
\usage{
microceph_v1(pars)
}
\arguments{
\item{pars}{the model parameters, specifying "mean" (gamma mean),"var" (gamma variance), and "c" (additional scaling constant)}
}
\value{
the vector of risks
}
\description{
Microcephaly risk curve under gamma distribution. 0 - 279 (all days)
}

