% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{microceph_v2}
\alias{microceph_v2}
\title{Microcephaly risk V2}
\usage{
microceph_v2(pars)
}
\arguments{
\item{pars}{the model parameters}
}
\value{
the vector of risks
}
\description{
Microcephaly risk curve with 3 distinct periods (14, 14 and 12 weeks)
}

