% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{microceph_v3}
\alias{microceph_v3}
\title{Microcephaly risk V3}
\usage{
microceph_v3(pars)
}
\arguments{
\item{pars}{the model parameters, "p1"-"p6"}
}
\value{
the vector of risks
}
\description{
Microcephaly risk curve with 6 distinct periods (7, 7, 7, 7, 7 and 5 weeks)
}

