% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{microceph_v4}
\alias{microceph_v4}
\title{Microcephaly risk V4}
\usage{
microceph_v4(pars)
}
\arguments{
\item{pars}{the model parameters, "p1"-"p8"}
}
\value{
the vector of risks
}
\description{
Microcephaly risk curve with 8 distinct periods
}

