% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{mvr_proposal}
\alias{mvr_proposal}
\title{Multivariate proposal function}
\usage{
mvr_proposal(current, param_table, covMat)
}
\arguments{
\item{current}{the vector of current parameter values}

\item{param_table}{a matrix of flags and bounds that specify how the proposal should be treated. Crucial elements are the 3rd column (lower bound), 4th column (upper bound), 5th column )step size) and 6th column (flag for log proposal)}

\item{covMat}{the 2D covariance matrix for all of the parameters}
}
\value{
a parameter vector of a proposed move. Note that these may fall outside the allowable ranges.
}
\description{
Given the current parameters, bounds and a covariance matrix, returns a vector for a proposed jump from a multivariate normal distribution
}

