% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_setup.R
\name{partab_setup}
\alias{partab_setup}
\title{Parameter setup}
\usage{
partab_setup(stateNames, version, realDat, useInc, allowablePars = NULL,
  sharedProb = TRUE, normLik = NULL, stateWeights = FALSE)
}
\arguments{
\item{stateNames}{the vector of states to include}

\item{version}{the version of the model to run (1-3)}

\item{realDat}{the data frame of actual microcephaly data. Need this to generate N_H and L_H}

\item{useInc}{bool indicating whether or not ZIKV incidence data is considered here}

\item{allowablePars}{data frame of allowable parameters for density and constSeed}

\item{sharedProb}{bool if the version with shared microcephaly parameters should be used}

\item{normLik}{Using normal likelihood function for data? If false, uses binomial likelihood. If true, uses normal likelihood which incorporates two standard deviation parameters for ZIKV and microcephaly incidence data}

\item{stateWeights}{bool deciding if data from different states should be weighted}
}
\value{
the modified parameter table
}
\description{
Sets up the parameter table based on the version, states and allowable parameters
}

