% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_setup.R
\name{parTabSetup}
\alias{parTabSetup}
\title{Parameter setup}
\usage{
parTabSetup(locationNames, version, locationInfo, useInc,
  allowablePars = NULL, sharedProb = TRUE, normLik = FALSE,
  locationWeights = FALSE, forecast = FALSE)
}
\arguments{
\item{locationNames}{the vector of locations to include}

\item{version}{Version 1-3 indicating which form of the microcephaly risk curve is to be used}

\item{locationInfo}{a table with location name, L_H (human lifespan in years) and N_H (human population size). See data(locationInfo)}

\item{useInc}{bool indicating whether or not ZIKV incidence data is considered here}

\item{allowablePars}{data frame of allowable parameters for density and t0}

\item{sharedProb}{bool if the version with shared microcephaly parameters should be used}

\item{normLik}{Using normal likelihood function for data. If false, uses binomial likelihood. If true, uses normal likelihood which incorporates two standard deviation parameters for ZIKV and microcephaly incidence data}

\item{locationWeights}{bool deciding if data from different locations should be weighted}

\item{forecast}{if TRUE, sets up the parameter table for the forecasting analysis}
}
\value{
the modified parameter table
}
\description{
Sets up the parameter table based on the version, locations and allowable parameters
}

