% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_setup.R
\name{parTabSetup_aux}
\alias{parTabSetup_aux}
\title{MCMC parameter setup}
\usage{
parTabSetup_aux(locationInfo, version = 1, sharedProb = FALSE,
  parFile = "", locationParFile = "", locationNormLik = FALSE,
  locationWeights = FALSE, forecast = FALSE)
}
\arguments{
\item{locationInfo}{data frame of location info, see data(locationInfo). Needs columns for L_H, N_H and rawName}

\item{version}{which model version is being run (see \code{\link{print_version_names}})}

\item{sharedProb}{boolean indicating if all locations share the same risk curve}

\item{parFile}{if already specified, where to find the parameter table file without location parameters}

\item{locationParFile}{as parFile, but with location specific parameters added}

\item{locationNormLik}{boolean deciding whether or not a normal likelihood function is used (otherwise binomial). If NULL, uses the binomial. If not NULL, true or false indicates whether the likelihood standard deviations are location specific or not.}

\item{locationWeights}{bool deciding if location weightings should be used (otherwise all locations have equal weighting)}

\item{forecast}{if TRUE, sets up the parameter table for the forecasting analysis}
}
\value{
a matrix of needed settings for the MCMC algorithm. For each parameter, gives a name, lower and upper bounds, boolean for log scale, initial step sizes, log proposal and whether or not the parameter should be fixed.
}
\description{
Sets up the parameter table for use in MCMC
}

