% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.dynamics}
\alias{plot.dynamics}
\title{SEIR dynamics plot}
\usage{
\method{plot}{dynamics}(y, N_H, N_M, file.name = NULL)
}
\arguments{
\item{y}{The data frame or matrix of times and population sizes}

\item{N_H}{human population size}

\item{N_M}{mosquito population size}

\item{file.name}{optional filename at which to save the plot. Must be a PNG. If NULL, does not open the png device.}
}
\description{
Plots SEIR dynamics given a data frame of solved ODEs
}

