% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{plot_best_pars}
\alias{plot_best_pars}
\title{Plot best pars}
\usage{
plot_best_pars(chain, dat, parTab, t_pars, local, number, runs = NULL)
}
\arguments{
\item{chain}{the MCMC chain}

\item{dat}{the data frame of real data}

\item{parTab}{the parameter table}

\item{t_pars}{time parameters}

\item{local}{string for state under consideration}

\item{number}{the number to be appended to parameter names in the MCMC chain}
}
\description{
Given an MCMC chain of fitting to the multiple data sets, plots the epidemic dynamics
}

