% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_best_trajectory_multi}
\alias{plot_best_trajectory_multi}
\title{Plot all best trajectoroes}
\usage{
plot_best_trajectory_multi(runName, chain = NULL, realDat = NULL,
  parTab = NULL, pars = NULL, runs = 100,
  runDir = "~/net/home/zika/outputs3", allDatFile = "allDat20.04.16.csv",
  nchains = 3, mcmcPars = c(burnin = 50000, adaptive = 1e+05, thin = 50))
}
\arguments{
\item{runName}{the type of model fitting done}

\item{runs}{the number of runs to use for prediction intervals}

\item{runDir}{the full directory path where the MCMC chains are located}

\item{allDatFile}{local filename of actual data OPTIONAL}

\item{state}{the name of the state to be fitted}

\item{ylabel}{boolean whether or not to add y axis label}

\item{xlabel}{boolean whether or not to add x axis label}

\item{nchain}{number of chains run}
}
\value{
a ggplot object with the incidence plots
}
\description{
For a given run name and state, plots the best zika and microcephaly incidence for all states
}

