% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_micro_bounds}
\alias{plot_micro_bounds}
\title{Plot microcephaly curves with bounds}
\usage{
plot_micro_bounds(chain, samp_no = 10000, ylim = 0.3, chain2 = NULL,
  colombia_chain = NULL)
}
\arguments{
\item{chain}{the MCMC chain}
}
\value{
the ggplot object
}
\description{
Given an MCMC chain, plots microcephaly risk curves with 95% prediction intervals
}

