% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_random_microceph_curves}
\alias{plot_random_microceph_curves}
\title{Plot microceph curves}
\usage{
plot_random_microceph_curves(chain, runs)
}
\arguments{
\item{chain}{The MCMC chain. Must have the gamma distribution mean, variance and scale}

\item{runs}{number of random draws to plot}
}
\value{
a ggplot object with the microcephaly risk curve
}
\description{
Given an MCMC chain, plots the best fitting microcephaly risk curve and some faint, random draws from the posterior
}

