% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_setup}
\alias{plot_setup}
\title{Auxiliary setup}
\usage{
plot_setup(runName, runDir = "~/net/home/zika/outputs3",
  allDatFile = "~/net/home/zika/allDat20.04.16.csv", nchains = 3,
  mcmcPars = c(burnin = 50000, adaptive = 1e+05, thin = 50))
}
\arguments{
\item{runName}{the name of the runto be plotted. This should match the filename of the MCMC chains}

\item{runDir}{the full directory path where the MCMC chains are located}

\item{allDatFile}{local filename of actual data OPTIONAL}

\item{mcmcPars}{named vector (burnin, adaptive, thin) to format the chain correctly}

\item{nchain}{number of chains run}
}
\value{
a list of all the parameters needed for plotting
}
\description{
Given the run name, sets up all of the parameters needed to produce all plots
}

