% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDataHeatMap}
\alias{plotDataHeatMap}
\title{Head circumference heatmap}
\usage{
plotDataHeatMap(dat)
}
\arguments{
\item{dat}{matrix of head count data. Rows represent sampling times and columns represent individual measurements.}
}
\value{
A ggplot object with the heatmap of head sizes over time. White line shows mean head size.
}
\description{
Given a matrix or data frame of head sizes over time (rows represent sampling times), plots a heatmap showing distribution and mean head sizes over time.
}

