% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior_apply}
\alias{posterior_apply}
\title{Oh wow}
\usage{
posterior_apply(place, t_pars, values, names, local, startDays, endDays,
  buckets, microCeph, births, data_locals, incDat = NULL, allPriors = NULL,
  peakTimes = NULL)
}
\arguments{
\item{ts}{time vector over which to solve the ODE model}

\item{y0s}{initial conditions for the ODE model}

\item{pars}{ODE parameters}

\item{dat}{unnamed matrix over which to calculate likelihoods}
}
\value{
a single value for the posterior
}
\description{
Given the time vector, initial conditions ODE parameters and a matrix of microcephaly data, calculates the posterior value for a given data set.
}

