% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior_complex}
\alias{posterior_complex}
\title{Posterior function for complex ODE model}
\usage{
posterior_complex(t_pars, y0s, pars, dat, threshold = NULL, times = NULL)
}
\arguments{
\item{y0s}{initial conditions for the ODE model}

\item{pars}{ODE parameters}

\item{dat}{unnamed matrix over which to calculate likelihoods}

\item{threshold}{optional paramter. If not null, should be the threshold for microcephaly diagnosis}

\item{times}{optional parameter. If not null, then this should be a matrix of times over which data is recorded}

\item{ts}{time vector over which to solve the ODE model}
}
\value{
a single value for the posterior
}
\description{
Given the time vector, initial conditions and ODE parameters, calculates the posterior value for a given data set. Note that no priors are used here.
}

