% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior_inc}
\alias{posterior_inc}
\title{Posterior function for the simple SEIR model with incidence only}
\usage{
posterior_inc(ts, values, names, local, inc_startDays, inc_endDays, inc_locals,
  inc_buckets, inc_ZIKV, inc_NH, unique_locations, allPriors = NULL,
  inc_valid_days = NULL, inc_valid_local = NULL)
}
\arguments{
\item{ts}{time vector over which to solve the ODE model}

\item{values}{model parameters}

\item{names}{names of the model parameters}

\item{local}{the vector of location names corresponding to the parameter table}

\item{inc_startDays}{vector of all starting days for reporting buckets}

\item{inc_endDays}{vector of all end days for reporting buckets}

\item{inc_buckets}{vector of all reporting bucket sizes}

\item{inc_ZIKV}{vector of reported ZIKV case}

\item{inc_NH}{vector of all population sizes}

\item{unique_locations}{vector of all unique locations to be tested}

\item{allPriors}{defaults to FALSE. Arguments for parameter priors, if desired.}

\item{inc_valid_days}{vector of all days included in the incidence data}

\item{inc_valid_local}{vector of locals corresponding to the days in inc_valid_days}

\item{inc_local}{vector of all location names corresponding to the incidence table}
}
\value{
a single value for the posterior
}
\description{
Given the time vector, ODE parameters and deconstructed matrix of incidence data, calculates the posterior probability for a given data set.
}

