% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior_known_inc_seir}
\alias{posterior_known_inc_seir}
\title{Posterior function for forecast}
\usage{
posterior_known_inc_seir(pars, startDays, endDays, buckets, microCeph, births,
  zikv, nh, inc_buckets, inc_start, inc_end, valid_days_micro, valid_days_inc)
}
\description{
Using the microcephaly risk model, produces a forecast of microcephaly affected births for all time periods and calculates a likelihood of observing given microcephaly cases
}

