% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{posterior_simple}
\alias{posterior_simple}
\title{Posterior function for the simple SEIR model}
\usage{
posterior_simple(t_pars, y0s, pars, dat)
}
\arguments{
\item{y0s}{initial conditions for the ODE model}

\item{pars}{ODE parameters}

\item{dat}{unnamed matrix over which to calculate likelihoods}

\item{ts}{time vector over which to solve the ODE model}
}
\value{
a single value for the posterior
}
\description{
Given the time vector, initial conditions ODE parameters and a matrix of microcephaly data, calculates the posterior value for a given data set. Note that no priors are used here.
}

