% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior_simple_buckets}
\alias{posterior_simple_buckets}
\title{Posterior function for the simple SEIR model with bucketed data for a single state}
\usage{
posterior_simple_buckets(ts, y0s, pars, startDays, endDays, buckets, microCeph,
  births, zikv = NULL, nh = NULL, inc_buckets = NULL, inc_start = NULL,
  inc_end = NULL, peak_start = NULL, peak_end = NULL, allPriors = NULL,
  place)
}
\arguments{
\item{ts}{time vector over which to solve the ODE model}

\item{y0s}{initial conditions for the ODE model}

\item{pars}{ODE parameters}

\item{startDays}{vector of the start day for each bucket of data}

\item{endDays}{vector of the end day for each bucket of data}

\item{buckets}{bucket sizes}

\item{microCeph}{vector of microcephaly incidence for each bucket}

\item{births}{vector of total briths for each bucket}

\item{zikv}{vector of ZIKV incidence}

\item{nh}{vector of population sizes}

\item{inc_buckets}{vector of bucket size for incidence data}

\item{inc_start}{vector of inc start days}

\item{inc_end}{vector of inc end days}

\item{peak_start}{vector of peak time start days}

\item{peak_end}{vector of peak time end days}

\item{allPriors}{defaults to FALSE. Arguments for parameter priors, if desired.}
}
\value{
a single value for the posterior
}
\description{
Given the time vector, initial conditions ODE parameters and a matrix of microcephaly data, calculates the posterior value for a given data set. Note that no priors are used here.
}

