% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{priors}
\alias{priors}
\title{All model priors}
\usage{
priors(pars)
}
\arguments{
\item{pars}{the vector of parameters}
}
\value{
a single log prior value
}
\description{
Takes the vector of model parameters and returns a single value for the log prior probability
}

