% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{proposalfunction}
\alias{proposalfunction}
\title{MCMC proposal function}
\usage{
proposalfunction(param, param_table, index)
}
\arguments{
\item{param}{a vector of the parameters to be explored}

\item{param_table}{a matrix of flags and bounds that specify how the proposal should be treated. Crucial elements are the 3rd column (lower bound), 4th column (upper bound), 5th column )step size) and 6th column (flag for log proposal)}

\item{index}{numeric value for the index of the parameter to be moved from the param table and vector}
}
\value{
the parameter vector after step
}
\description{
Proposal function for MCMC random walk, taking random steps of a given size. Random walk may be on a linear or log scale
}

