% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{r0.vector}
\alias{r0.vector}
\title{R0 vector calculation}
\usage{
r0.vector(params)
}
\arguments{
\item{params}{Matrix of parameters matching those returned by \code{\link{setupListPars}}}
}
\value{
A vector of values for R0
}
\description{
Calculates the R0 of the SEIR model given a matrix of parameters. R0 defined as number of expected human cases given introduction of 1 infected human into a totally naive population of humans and mosquitoes.
}
\seealso{
\code{\link{b.calc}}
}

