% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{read_all_data}
\alias{read_all_data}
\title{Read all data}
\usage{
read_all_data(microCephFile = "~/net/home/zika/data/microcephaly_data.csv",
  incDatFile = "~/net/home/zika/data/inc_data.csv", epidemicStates = FALSE,
  order = TRUE, convert = TRUE)
}
\arguments{
\item{microCephFile}{full filepath to microcephaly data}

\item{incDatFile}{full filepath ot incidence data}

\item{epidemicStates}{if TRUE, gets only epidemic states. Otherwise gives all states}

\item{order}{if TRUE, orders the states by per capita microcephaly rate}

\item{convert}{if TRUE, converts strings to factors}
}
\value{
a list of microcephaly and incidence data
}
\description{
Function to read in microcephaly and incidence data. Converts strings to factors, and orders them by microcephaly cases
}

