% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{read_inipars}
\alias{read_inipars}
\title{Read partab}
\usage{
read_inipars(location = "")
}
\arguments{
\item{location}{the path to the directory to read from}
}
\value{
the full parameter table
}
\description{
Read in the initial parameter file, which is the same as the parameter table produced from zikaInfer
}

