% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD.R
\name{rm_scale}
\alias{rm_scale}
\title{Step size scaler}
\usage{
rm_scale(step_scale, mc, popt, log_prob, N_adapt)
}
\arguments{
\item{step_scale}{the current step size}

\item{mc}{the number of MCMC iterations run}

\item{popt}{the desired acceptance rate}

\item{log_prob}{the current log probability}

\item{N_adapt}{the total number of adaptive steps to be taken}
}
\description{
Given current step scale, desired optimum step size etc, adapts the MCMC step size
}

