% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{sampler_slice}
\alias{sampler_slice}
\title{Slice sampler}
\usage{
sampler_slice(lik, x_init, fixed, y_init, w, lower, upper)
}
\arguments{
\item{lik}{target function}

\item{x_init}{intial parameter values for the target function}

\item{fixed}{vector of bools to inidicate which parameters should be updated}

\item{y_init}{current target function value}

\item{w}{vector of widths for the slice sampler}

\item{lower}{optional vector of lower parameter bounds}

\item{upper}{options vector of upper parameter points}
}
\value{
a list. First component is updated parameters, second is corresponding target function evaluation
}
\description{
Multivariate slice sampler for given target function
}

