% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setupListPars}
\alias{setupListPars}
\title{Default parameter list setup}
\usage{
setupListPars(duration = 10)
}
\arguments{
\item{duration}{Duration in years of the simulation. Defaults to 10 years}
}
\value{
list of times, initial conditions, and ODE parameters
}
\description{
Sets up the default list of parameters and initial conditions to be used by \code{\link{zika.sim}}
}
\seealso{
\code{\link{setupParsLong}}
}

