% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_setup.R
\name{setupLocationParTable}
\alias{setupLocationParTable}
\title{Setup location specific parameter table}
\usage{
setupLocationParTable(locationDat, locationParFile = "locationParams.csv")
}
\arguments{
\item{locationDat}{the data frame of all data}

\item{locationParFile}{file name for the parameter table if exists}
}
\value{
a parameter table
}
\description{
Takes the entire data set and returns a useable format parameter table for the place specific parameters
}

