% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setupParsODE}
\alias{setupParsODE}
\title{Defaults ODE parameters}
\usage{
setupParsODE()
}
\value{
vector of named parameter values
}
\description{
Generates the default vector of parameters for the ODE model
}

