% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setupStateParTable}
\alias{setupStateParTable}
\title{Setup state specific parameter table}
\usage{
setupStateParTable(stateDat, stateParFile = "stateParams.csv")
}
\arguments{
\item{stateDat}{the data frame of all data}

\item{stateParFile}{file name for the parameter table if exists}
}
\value{
a parameter table
}
\description{
Takes the entire data set and returns a useable format parameter table for the place specific parameters
}

