% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{simeq}
\alias{simeq}
\title{Attack rate simeq}
\usage{
simeq(par, R0)
}
\arguments{
\item{par}{the current estimated attack rate}

\item{R0}{the value of R0 to be tested}
}
\value{
the difference between the two sides of the final size calculation
}
\description{
Simeq cost function for calculating attack rate from R0
}

