% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_1d}
\alias{slice_1d}
\title{Univariate slice sampler}
\usage{
slice_1d(f, x_init, y_init, w, lower, upper)
}
\arguments{
\item{f}{target function}

\item{x_init}{intial parameter value for the target function}

\item{y_init}{current target function value}

\item{w}{width for the slice sampler}

\item{lower}{lower parameter bounds}

\item{upper}{upper parameter points}
}
\value{
a single sampled value
}
\description{
Univariate slice sampler for given target function
}

