% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{solveModel}
\alias{solveModel}
\title{Solve ODE model}
\usage{
solveModel(allPars)
}
\arguments{
\item{allPars}{the list of parameters needed for the ODE model}
}
\value{
a data frame of the solved ODE model
}
\description{
Given a list of parameters as generated by \code{\link{setupListPars}}, solves the ODE model and returns a named data frame that can be passed to \code{\link{plot_dynamics}}
}
\seealso{
\code{\link{solveModelSimple}}
}

