% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{solveModelComplex}
\alias{solveModelComplex}
\title{Solve ODE model with age classes}
\usage{
solveModelComplex(t_pars, y0s, pars)
}
\arguments{
\item{t_pars}{vector of two components - the total run time in days and the time step for the ODE solver}

\item{y0s}{the initial population sizes for the ODE model}

\item{pars}{a named vector with all of the necessary parameters to solve the ODE model. See \code{\link{setupParsODE}}}
}
\value{
a data frame of the solved ODE model
}
\description{
Given a list of parameters as generated by \code{\link{setupListPars}}, solves the ODE model and returns a named data frame that can be passed to \code{\link{plot_dynamics}}
}
\seealso{
\code{\link{solveModelSimple}}
}

