% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{solveModelSimple}
\alias{solveModelSimple}
\title{Solve simple SEIR model}
\usage{
solveModelSimple(t_pars, y0s, pars)
}
\arguments{
\item{t_pars}{vector of two components - the total run time in days and the time step for the ODE solver}

\item{y0s}{the initial population sizes for the ODE model}

\item{pars}{a named vector with all of the necessary parameters to solve the ODE model. See \code{\link{setupParsODE}}}
}
\value{
a data frame of the solved ODE model
}
\description{
Given a list of parameters as generated by \code{\link{setupListPars}}, solves the simple SEIR model and returns a named data frame
}
\seealso{
\code{\link{solveModel}}
}

