% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{solveModelSimple_rlsoda}
\alias{solveModelSimple_rlsoda}
\title{Solve simple SEIR model}
\usage{
solveModelSimple_rlsoda(ts, y0s, pars, compatible = FALSE)
}
\arguments{
\item{ts}{vector of times to solve the ODE model over}

\item{y0s}{the initial population sizes for the ODE model}

\item{pars}{a named vector with all of the necessary parameters to solve the ODE model. See \code{\link{setupParsODE}}}

\item{compatible}{indicates if solved data frame should be given in deSolve return format}
}
\value{
a data frame of the solved ODE model
}
\description{
Given a list of parameters as generated by \code{\link{setupListPars}}, solves the simple SEIR model and returns a named data frame. Uses rlsoda from rlsoda
}

