% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{solveSEIRModel}
\alias{solveSEIRModel}
\title{Solve simple SEIR model}
\usage{
solveSEIRModel(ts, y0s, pars, solver = "rlsoda", compatible = FALSE)
}
\arguments{
\item{ts}{vector of times to solve the ODE model over}

\item{y0s}{the initial population sizes for the ODE model}

\item{pars}{a named vector with all of the necessary parameters to solve the ODE model. See \code{\link{setupParsODE}}}

\item{solver}{which ODE solver to use, either "rlsoda" or "lsoda"}

\item{compatible}{if TRUE, makes the rlsoda output the same as lsoda output}
}
\value{
a data frame of the solved ODE model
}
\description{
Solves the SEIR model using the specified ode solver
}

