% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{solveSEIRModel_lsoda}
\alias{solveSEIRModel_lsoda}
\title{Solve simple SEIR model (lsoda)}
\usage{
solveSEIRModel_lsoda(ts, y0s, pars, makenames = FALSE)
}
\arguments{
\item{ts}{vector of times to solve the ODE model over}

\item{y0s}{the initial population sizes for the ODE model}

\item{pars}{a named vector with all of the necessary parameters to solve the ODE model. See \code{\link{setupParsODE}}}

\item{makenames}{indicates if solved data frame should be given column names}
}
\value{
a data frame of the solved ODE model
}
\description{
Given a list of parameters as generated by \code{\link{setupListPars}}, solves the simple SEIR model and returns a named data frame. Uses lsoda from ODE
}

