% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{states_to_title}
\alias{states_to_title}
\title{Get states uppercase}
\usage{
states_to_title()
}
\value{
vector of all states
}
\description{
Gives a vector of all states in Brazil. Names are lowercase names, values are actual names
}

