% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sum_n_vector}
\alias{sum_n_vector}
\title{Sums nth elements}
\usage{
sum_n_vector(v, n)
}
\arguments{
\item{v}{the vector to me summed}

\item{n}{the number of elements to sum}
}
\description{
Sums every n values of a numeric vector
}

