% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{summarise_chain}
\alias{summarise_chain}
\title{Summarise MCMC chain}
\usage{
summarise_chain(chain)
}
\arguments{
\item{chain}{the MCMC chain}
}
\value{
a data frame of MCMC summary stats
}
\description{
Gives the summary statistics and quantiles of all columns from an MCMC chain
}

