% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{un_cumsum}
\alias{un_cumsum}
\title{Un cumulative sum}
\usage{
un_cumsum(x)
}
\arguments{
\item{x}{the vector to be "uncumsummed"}
}
\value{
the un cumulative summed vector
}
\description{
Given a vector representing a cumulative sum, returns a vector of the same length with new "incidence" for each time point
}

