% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcano_plots}
\alias{volcano_plots}
\title{Volcano plots}
\usage{
volcano_plots(chains, parTab)
}
\arguments{
\item{chains}{the list of chains}

\item{parTab}{the parameter table used for correct indexing}
}
\value{
a ggplot table
}
\description{
Given a list of MCMC chains, produces a grid of volcano plots for the unfixed state-specific parameters
}

