% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{zika.ode}
\alias{zika.ode}
\title{Obsolete - the R implementation of the ODE model}
\usage{
zika.ode(t, y, pars)
}
\arguments{
\item{t}{the vector of times to be solved over}

\item{y}{the compartment states}

\item{pars}{the parameters used to solve the ODE model}
}
\value{
set of derivatives at the given time point
}
\description{
Set of ODEs for the zika SEIR model to be used by deSolve
}

