% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{zika.sim}
\alias{zika.sim}
\title{Simulates head circumference data from an SEIR vector borne model}
\usage{
zika.sim(allPars)
}
\arguments{
\item{allPars}{A list of parameter vectors. First vector should be an array of time points over which to solve the SEIR model. Second vector should be the vector of starting conditions. Third vector should be all of the parameters.}
}
\value{
a matrix of simulated head circumferences over time.
}
\description{
Takes a list of necessary arguments to solve a system of ODEs that generates head circumferences of new born infants
}

