% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{calculate_BIC}
\alias{calculate_BIC}
\title{BIC calculation}
\usage{
calculate_BIC(chain, state, parTab, allDat, incDat)
}
\arguments{
\item{chain}{the MCMC chain to be tested}

\item{state}{the subset state to take. Note that 'local' must be a column to subset rows by.}

\item{parTab}{the parameter table used for this chain}

\item{allDat}{the microcephaly data}

\item{incDat}{the incidence data}
}
\value{
a single BIC value
}
\description{
Given an MCMC chain melted by state and subset state, calculates the BIC for the given chain
}

