% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_posterior}
\alias{create_posterior}
\title{Simplify posterior call}
\usage{
create_posterior(ts, values, names, local, startDays, endDays, buckets,
  microCeph, births, data_locals, inc_startDays = NULL, inc_endDays = NULL,
  inc_locals = NULL, inc_buckets = NULL, inc_ZIKV = NULL, inc_NH = NULL,
  peak_startDays = NULL, peak_endDays = NULL, peak_locals = NULL,
  unique_locations, allPriors = NULL, version = "normal")
}
\arguments{
\item{ts}{time vector over which to solve the ODE model}

\item{values}{ODE parameters}

\item{names}{names for the ODE parameters}

\item{local}{names of all the locations that are being considered}

\item{startDays}{vector of the start day for each bucket of data}

\item{endDays}{vector of the end day for each bucket of data}

\item{buckets}{bucket sizes}

\item{microCeph}{vector of microcephaly incidence for each bucket}

\item{births}{vector of total briths for each bucket}

\item{data_locals}{corresponding vector of location names for the microcephaly data}

\item{inc_startDays}{start day vector for incidence data}

\item{inc_endDays}{end day vector for incidence data}

\item{inc_locals}{vector of location names for incidence data}

\item{inc_buckets}{sampling window vector for incidence data}

\item{inc_ZIKV}{vector of ZIKV incidence for each time point}

\item{inc_NH}{vector of population size over time}

\item{peak_startDays}{vector of initial allowable day of incidence peak time}

\item{peak_endDays}{vector of end allowable day of incidence peak time}

\item{peak_locals}{vector of corresponding location name}

\item{unique_locations}{vector of all unique location names}

\item{allPriors}{defaults to NULL. Arguments for parameter priors, if desired.}

\item{version}{usually just leave this as "normal". I've added a "forecast" version which expects parameters to do with the lack of second wave.}
}
\value{
a single value for the posterior
}
\description{
Simplifies the call to the posterior function using closures
}

