% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_setup.R
\name{createParTable}
\alias{createParTable}
\title{MCMC parameter table creation}
\usage{
createParTable(saveFile = NULL)
}
\arguments{
\item{saveFile}{if provided, writes the parameter table to the given filename}
}
\value{
a matrix of needed settings for the MCMC algorithm. For each parameter, gives a name, lower and upper bounds, initial step sizes and whether or not the parameter should be fixed.
}
\description{
Sets up the parameter table for use in MCMC
}

