% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{exampleParTab}
\alias{exampleParTab}
\title{Example parameter table}
\format{A data frame}
\usage{
data(exampleParTab)
}
\description{
An example of parameter table that would be used to generate simulated data and control parameter space exploration during the MCMC procedure.
}
\keyword{datasets}

