% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{extra_microceph_calculations}
\alias{extra_microceph_calculations}
\title{Post calculations}
\usage{
extra_microceph_calculations(chain, version = 1, microceph_limit = 0.001,
  scale = 1)
}
\arguments{
\item{chain}{the full MCMC chain to do post analysis on}

\item{version}{the version of the model run here}

\item{microceph_limit}{the probability of microcephaly given infection to use as the cut off}

\item{scale}{if a non 100 percent reporting proportion, need to divide by assumed proportion}
}
\value{
an MCMC chain of the added analyses
}
\description{
Adds some parameter estimates for peripheral microcephaly risk parameters eg. mode. Note that this only applies to version 1 of the model. If not version 1, will return lots of zeros
}

