% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_micro_curve}
\alias{generate_micro_curve}
\title{Generate microceph curve}
\usage{
generate_micro_curve(pars)
}
\arguments{
\item{pars}{the model parameters}
}
\value{
a vector of risks
}
\description{
Generates the microcephaly risk curve from model parameters
}

