% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_microceph_dat}
\alias{generate_microceph_dat}
\title{Simulation microcephaly data}
\usage{
generate_microceph_dat(t_pars, y0s, pars, births, weeks = FALSE)
}
\arguments{
\item{t_pars}{vector of two components - the total run time in days and the time step for the ODE solver}

\item{y0s}{the initial population sizes for the ODE model}

\item{pars}{a named vector with all of the necessary parameters to solve the ODE model. See \code{\link{setupParsODE}}}

\item{births}{number of actual births per month}

\item{weeks}{Defaults to FALSE. If true. returns weekly microcephaly births rather than by month.}
}
\value{
a data frame of microcephaly births, total births and corresponding times.
}
\description{
Given all of the parameters needed to solve the SEIR model, produces a data frame of simulated microcephaly affected births for a single state
}

