% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_data.R
\name{get_correct_order}
\alias{get_correct_order}
\title{Location order}
\usage{
get_correct_order(named = TRUE, normalised = FALSE,
  filePath = "~/net/home/zika/data/microcephaly_data.csv")
}
\arguments{
\item{named}{if TRUE, converts the location names to full names}

\item{normalised}{if TRUE, orders by per birth microcephaly rate}

\item{filePath}{the full file locatoin of the microcephaly data}

\item{birth_dat}{the data frame of data}
}
\value{
a vector of ordered location names
}
\description{
Gets the correct location order for all plots, sorted by number of microcephaly cases
}

