% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analyses.R
\name{get_lower_micro_bound}
\alias{get_lower_micro_bound}
\title{Lower microbound}
\usage{
get_lower_micro_bound(chain, limit = 0.001)
}
\arguments{
\item{chain}{the MCMC chain to run over}

\item{limit}{the probability of microcephaly considered at risk}
}
\value{
a vector of first risk days
}
\description{
For a given MCMC chain, calculates the first day at risk
}
\seealso{
\code{\link{get_microceph_range}}
}

