% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{incidence_likelihood}
\alias{incidence_likelihood}
\title{Incidence likelihood}
\usage{
incidence_likelihood(perCapInc, inc, N_H)
}
\arguments{
\item{perCapInc}{expected per capita incidence of Zika}

\item{dat}{two column matrix. First column is observed Zika cases, second is total possible zika cases}
}
\value{
a single log likelihood
}
\description{
Given time-varying probabilities of observing Zika incidence and actual data, returns a single likelihood
}

