% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_best_trajectory_multi}
\alias{plot_best_trajectory_multi}
\title{Plot all best trajectoroes}
\usage{
plot_best_trajectory_multi(chain, realDat, parTab, t_pars, runs = 100,
  incDat = NULL, mcmcPars = c(burnin = 50000, adaptive = 1e+05, thin = 50),
  ylimM = 200, ylimI = 0.01)
}
\arguments{
\item{chain}{the MCMC chain of estimated parameters}

\item{realDat}{the data frame of data used for fitting}

\item{parTab}{the parameter table used for fitting}

\item{t_pars}{the time parameters used for solving the ODEs}

\item{runs}{the number of runs to use for prediction intervals}

\item{incDat}{optional data frame of actual incidence data}

\item{mcmcPars}{a named vector of the burnin, adaptive period and thinning}

\item{ylimM}{ylimit for the microcephaly plot}

\item{ylimI}{ylimit for the incidence plot}
}
\value{
a ggplot object with the incidence plots
}
\description{
For a given run name and state, plots the best zika and microcephaly incidence for all states
}

