% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD.R
\name{plot_best_trajectory_single}
\alias{plot_best_trajectory_single}
\title{Plot single best trajectory}
\usage{
plot_best_trajectory_single(local, chain = NULL, realDat = NULL,
  parTab = NULL, ts = NULL, runs = 100, incDat = NULL, ylabel = TRUE,
  xlabel = TRUE, mcmcPars = c(burnin = 50000, adaptive_period = 1e+05, thin
  = 50), ylimM = NULL, ylimI = NULL, startDay = NULL, months = NULL,
  weeks = NULL, xlim = NULL)
}
\arguments{
\item{local}{the name of the state to be fitted}

\item{chain}{the MCMC chain to use}

\item{realDat}{the data frame of real data that was fitted to}

\item{parTab}{the parameter table as returned by \code{\link{setupParTable}}}

\item{ts}{vector of time parameters}

\item{runs}{the number of runs to use for prediction intervals}

\item{incDat}{incidence data set, if to be included in the plot}

\item{ylabel}{boolean whether or not to add y axis label}

\item{xlabel}{boolean whether or not to add x axis label}

\item{mcmcPars}{a named vector of the burnin, adaptive period and thinning}

\item{ylimM}{ylimit for the microcephaly plot}

\item{ylimI}{ylimit for the incidence plot}

\item{startDay}{if realDat is not provided, need to provide the first day on which we want to see predicted microcephaly numbers}

\item{months}{if realDat not provided, need to provide the number of months of forecasted data that we wish to see}

\item{weeks}{if no incidence data provided, number of weeks over which we should plot the data}

\item{xlim}{OPTIONAL x coordinate limits}
}
\value{
a ggplot object with the incidence plots
}
\description{
For a given run name and state, plots the best zika and microcephaly incidence plot
}

