% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_best_trajectory_single}
\alias{plot_best_trajectory_single}
\title{Plot single best trajectory}
\usage{
plot_best_trajectory_single(local, chain = NULL, realDat = NULL,
  parTab = NULL, t_pars = NULL, runs = 100, incDat = NULL,
  ylabel = TRUE, xlabel = TRUE, mcmcPars = c(burnin = 50000,
  adaptive_period = 1e+05, thin = 50), ylimM = 200, ylimI = 0.02)
}
\arguments{
\item{runs}{the number of runs to use for prediction intervals}

\item{incDat}{optional data frame of actual incidence data}

\item{ylabel}{boolean whether or not to add y axis label}

\item{xlabel}{boolean whether or not to add x axis label}

\item{mcmcPars}{a named vector of the burnin, adaptive period and thinning}

\item{ylimM}{ylimit for the microcephaly plot}

\item{ylimI}{ylimit for the incidence plot}

\item{runName}{the type of model fitting done}

\item{state}{the name of the state to be fitted}
}
\value{
a ggplot object with the incidence plots
}
\description{
For a given run name and state, plots the best zika and microcephaly incidence plot
}

