% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_MCMC}
\alias{plot_MCMC}
\title{Plot MCMC chains}
\usage{
plot_MCMC(chain, parTab = NULL)
}
\arguments{
\item{parTab}{optional parameter table}

\item{chains}{list of MCMC chains}
}
\value{
nothing - saves plots to pdf
}
\description{
Given an MCMC chain and optionally the parameter table, plots the MCMC chains and densities using coda
}

