% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_MCMC}
\alias{plot_MCMC}
\title{Plot MCMC chains}
\usage{
plot_MCMC(chains, parTab = NULL, filename = "mcmcplots.pdf")
}
\arguments{
\item{chains}{list of MCMC chains}

\item{parTab}{optional parameter table}

\item{filename}{the name of the file to save to}
}
\value{
nothing - saves plots to pdf
}
\description{
Given a list of MCMC chains and optionally the parameter table, plots the MCMC chains and densities using coda
}

