% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_setup_data}
\alias{plot_setup_data}
\title{Get plot data}
\usage{
plot_setup_data(chain, dat, parTab, ts, local, runs = NULL)
}
\arguments{
\item{chain}{the MCMC chain}

\item{dat}{the data frame of real data}

\item{parTab}{the parameter table}

\item{ts}{vector of times}

\item{local}{string for state under consideration}

\item{runs}{the number of samples to take}
}
\description{
Given an MCMC chain and other parameters, returns a list of data frames that can be used to generate an incidence plot
}

