% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{posterior}
\alias{posterior}
\title{Posterior function for ODE model}
\usage{
posterior(ts, y0s, pars, dat)
}
\arguments{
\item{ts}{time vector over which to solve the ODE model}

\item{y0s}{initial conditions for the ODE model}

\item{pars}{ODE parameters}

\item{dat}{unnamed matrix over which to calculate likelihoods}
}
\value{
a single value for the posterior
}
\description{
Given the time vector, initial conditions and ODE parameters, calculates the posterior value for a given data set. Note that no priors are used here.
}

