% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior}
\alias{posterior}
\title{Posterior function for ODE model}
\usage{
posterior(t_pars, y0s, values, names, local, startDays, endDays, buckets,
  microCeph, births, data_locals, version = 1, threshold = NULL,
  times = NULL, incDat = NULL, allPriors = NULL, actualPeakTimes = NULL)
}
\arguments{
\item{y0s}{initial conditions for the ODE model}

\item{version}{optional parameter to specify which model we are solving. If 1, this is the simple SEIR model}

\item{threshold}{optional paramter. If not null, should be the threshold for microcephaly diagnosis}

\item{times}{optional parameter. If not null, then this should be a matrix of times over which data is recorded}

\item{ts}{time vector over which to solve the ODE model}

\item{pars}{ODE parameters}

\item{dat}{unnamed matrix over which to calculate likelihoods}
}
\value{
a single value for the posterior
}
\description{
Given the time vector, initial conditions and ODE parameters, calculates the posterior value for a given data set. Note that no priors are used here.
}

