% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior_complex_buckets}
\alias{posterior_complex_buckets}
\title{Posterior function for the simple SEIR model with bucketed data for multiple states}
\usage{
posterior_complex_buckets(t_pars, values, names, local, startDays, endDays,
  buckets, microCeph, births, data_locals, incDat = NULL, peakTimes = NULL,
  allPriors = NULL)
}
\arguments{
\item{names}{names for the ODE parameters}

\item{local}{names of all the states that are being considered}

\item{startDays}{vector of the start day for each bucket of data}

\item{endDays}{vector of the end day for each bucket of data}

\item{buckets}{bucket sizes}

\item{microCeph}{vector of microcephaly incidence for each bucket}

\item{births}{vector of total briths for each bucket}

\item{data_locals}{corresponding vector of state names for the microcephaly data}

\item{incDat}{defaults to NULL. If provided, uses the incidence data to improve parameter estimates.}

\item{allPriors}{defaults to NULL. Arguments for parameter priors, if desired.}

\item{ts}{time vector over which to solve the ODE model}

\item{pars}{ODE parameters}

\item{actualPeakTimes}{defaults to NULL. Data frame of priors belief on incidence peak times. Provide a row for each state and three columns - the name of the state, the lower bound and the upper bound on peak time. Puts a uniform prior on this range.}
}
\value{
a single value for the posterior
}
\description{
Given the time vector, initial conditions ODE parameters and a matrix of microcephaly data, calculates the posterior value for a given data set.
}

