% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior_simple_buckets}
\alias{posterior_simple_buckets}
\title{Posterior function for the simple SEIR model with bucketed data}
\usage{
posterior_simple_buckets(t_pars, y0s, pars, startDays, endDays, buckets,
  microCeph, births, incDat = NULL, allPriors = NULL,
  actualPeakTime = NULL)
}
\arguments{
\item{y0s}{initial conditions for the ODE model}

\item{pars}{ODE parameters}

\item{ts}{time vector over which to solve the ODE model}

\item{dat}{unnamed matrix over which to calculate likelihoods}
}
\value{
a single value for the posterior
}
\description{
Given the time vector, initial conditions ODE parameters and a matrix of microcephaly data, calculates the posterior value for a given data set. Note that no priors are used here.
}

