% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setupListPars}
\alias{setupListPars}
\title{Default parameter list setup}
\usage{
setupListPars(duration = 2 * 365, N_H = 1e+06, N_M = 3 * N_H,
  version = 1)
}
\arguments{
\item{duration}{Duration in years of the simulation. Defaults to 10 years}
}
\value{
list of times, initial conditions, and ODE parameters
}
\description{
Sets up the default list of parameters and initial conditions to be used by \code{\link{zika.sim}}. Can specify which version of the model to use. Passing 1 gives the simple model with no age classes, otherwise gives the parameters for the model with age classes.
}
\seealso{
\code{\link{setupParsLong}}
}

