% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setupParsLong}
\alias{setupParsLong}
\title{Default simulation parameters}
\usage{
setupParsLong()
}
\value{
vector of names parameter values
}
\description{
Generates the default vector of parameters for the simulation
}
\seealso{
\code{\link{setupParsODE}}
}

