% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_setup.R
\name{setupParsODE}
\alias{setupParsODE}
\title{Defaults ODE parameters}
\usage{
setupParsODE(version = 1)
}
\arguments{
\item{version}{the version of the model to be used (see \code{\link{print_version_names}})}
}
\value{
vector of named parameter values
}
\description{
Generates the default vector of parameters for the ODE model. Can specify which version of the model to use. Passing 1 gives the simple model with no age classes, otherwise gives the parameters for the model with age classes.
}

