% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{setupParTable}
\alias{setupParTable}
\title{MCMC parameter setup}
\usage{
setupParTable(version = 1)
}
\arguments{
\item{pars}{the vector of parameters that would be used for the ODE model}
}
\value{
a matrix of needed settings for the MCMC algorithm. For each parameter, gives a name, lower and upper bounds, boolean for log scale, initial step sizes, log proposal and whether or not the parameter should be fixed.
}
\description{
Sets up the parameter table for use in \code{\link{run_metropolis_MCMC}}
}

