% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{simeq}
\alias{simeq}
\title{Final size calculation}
\usage{
simeq(par, R0)
}
\arguments{
\item{par}{array of length one containing the attack rate}

\item{R0}{the R0 used for solving}
}
\value{
difference between the LHS and RHS of the final size equation
}
\description{
Cost function to be used for the final epidemic size calculation. Using a given R0, if the argument of par[1] is the correct attack rate then this function will return 0. Try: nleqslv(A0, simeq, R0=3)
}

