% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{solveModelSimple}
\alias{solveModelSimple}
\title{Solve simple SEIR model}
\usage{
solveModelSimple(t_pars, y0s, pars)
}
\arguments{
\item{allPars}{the list of parameters needed for the ODE model}
}
\value{
a data frame of the solved ODE model
}
\description{
Given a list of parameters as generated by \code{\link{setupListPars}}, solves the simple SEIR model and returns a named data frame
}
\seealso{
\code{\link{solveModel}}
}

